virtual/pam: Drop outdated package
The virtual/pam package provided a dependency on either pam or
openpam. It looks like openpam was dropped from Gentoo rendering the
virtual/pam package unnecessary. And as such, it was also dropped from
Gentoo, and so we follow suit. It also helps updating sys-libs/pam (in
coreos-overlay) to a newer version by not generating some dependency
cycles.